fourth try

+++ src/Engineer.cpp
@@ -9,7 +9,7 @@ void Engineer::setProject(Project *project) {
 }
 void Engineer::calc() {
   double payment = calcBudgetPart(
-      1 / (this->project->getNumOfWorkers()), this->project->getBudget())
+      (double)1 / (this->project->getNumOfWorkers()), this->project->getBudget())
       + (this->getSalary() * this->getWorkTime())
       + this->getAddPayment();
   this->setPayment(payment);
